chore: move `Bytecode`, `Requests` and `Withdrawals` to `reth-primitives-traits` (#8954)
