+++ crates/primitives-traits/Cargo.toml
@@ -14,7 +14,7 @@ workspace = true
 [dependencies]
 reth-codecs.workspace = true
 
-alloy-consensus.workspace = true
+alloy-consensus = { workspace = true, features = ["serde"] }
 alloy-eips.workspace = true
 alloy-genesis.workspace = true
 alloy-primitives.workspace = true
@@ -22,11 +22,12 @@ alloy-rlp.workspace = true
 alloy-rpc-types-eth = { workspace = true, optional = true } 
 
 derive_more.workspace = true
-revm-primitives.workspace = true
+revm-primitives = { workspace = true, features = ["serde"] }
 
 # misc
 thiserror-no-std = { workspace = true, default-features = false }
 roaring = "0.10.2"
+byteorder = "1"
 
 # required by reth-codecs
 modular-bitfield.workspace = true
@@ -51,6 +52,7 @@ default = ["std"]
 std = ["thiserror-no-std/std"]
 test-utils = ["arbitrary"]
 arbitrary = [
+    "alloy-consensus/arbitrary",
     "dep:arbitrary",
     "dep:proptest",
     "dep:proptest-derive"
